+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/HDRenderPipelineAsset.Migration.cs
@@ -20,6 +20,10 @@ namespace UnityEngine.Rendering.HighDefinition
             ScalableSettingsRefactor,
             ShadowFilteringVeryHighQualityRemoval,
             SeparateColorGradingAndTonemappingFrameSettings,
+
+            // Note: This must be updated each time we add a new version to the enum.
+            // The Last version is used when we create a new HDRP asset so it does not migrate.
+            Last = SeparateColorGradingAndTonemappingFrameSettings,
         }
 
         static readonly MigrationDescription<Version, HDRenderPipelineAsset> k_Migration = MigrationDescription.New(
@@ -82,7 +86,7 @@ namespace UnityEngine.Rendering.HighDefinition
         );
 
         [SerializeField]
-        Version m_Version;
+        Version m_Version = Version.Last;
         Version IVersionable<Version>.version { get => m_Version; set => m_Version = value; }
 
         void Awake() => k_Migration.Migrate(this);